refactor: 문의 기능 Context 제거 및 API 기반 구조로 개선
